Scrubbing JWA/HDMB references (still)

+++ app/controllers/collections_controller.php
@@ -52,7 +52,7 @@ class CollectionsController extends Kea_Action_Controller
 	{
 		if( $id = self::$_request->getProperty( 'collection_id' ) ) {
 			/**
-			 * Nested functionality (HDMB)
+			 * Nested functionality
 			 * Rearrange order so that nested child collections have the collection_parent of the current collection
 			 */
 			$collections = $this->_findChildren($id);
@@ -62,14 +62,22 @@ class CollectionsController extends Kea_Action_Controller
 				$collection->save();
 			}
 			
-			$deleteObjects = self::$_request->getProperty( 'delete_objects' );
-			if($deleteObjects)
+			/**
+			 * This will delete all the objects in the collection if the right $_request variable is set,
+			 * for obvious reasons only a super user should be able to do this.
+			 *
+			 */
+			if( self::$_session->isSuper() )
 			{
-				$objMapper = new Object_Mapper;
-				$objects = $objMapper->find()->where('collection_id = ?', $id)->execute();
-				foreach($objects as $object)
+				$deleteObjects = self::$_request->getProperty( 'delete_objects' );
+				if($deleteObjects)
 				{
-					$object->delete();
+					$objMapper = new Object_Mapper;
+					$objects = $objMapper->find()->where('collection_id = ?', $id)->execute();
+					foreach($objects as $object)
+					{
+						$object->delete();
+					}
 				}
 			}
 			
@@ -114,7 +122,12 @@ class CollectionsController extends Kea_Action_Controller
 		return false;
 	}
 	
-	//HDMB specific for nested functionality
+	/**
+	 * Find the children of the current collection (this only works if collection_parent is implemented)
+	 *
+	 * @return void
+	 * @author Kris Kelly
+	 **/
 	protected function _findChildren($parent = NULL)
 	{
 		$mapper = new Collection_Mapper();
@@ -152,7 +165,13 @@ class CollectionsController extends Kea_Action_Controller
 		return $mapper->addToCollection( $obj_id, $coll_id );
 	}
 	
-	//HDMB specific
+	/**
+	 * This is a total hack to display nested collections all at once on the same page using recursive partials.
+	 * It would be better just to use findChildren() and deprecate this.  See content/admin/collections/all.php
+	 *
+	 * @return void
+	 * @author Kris Kelly
+	 **/
 	public function displayNested($template, $useList = TRUE, $style = NULL, $parent_id = NULL)
 	{
 		$mapper = new Collection_Mapper;
